pyproject.toml
[project]
name = "amis-agent"
version = "0.1.0"
description = "Digital Me Agent for automated lead discovery and outreach"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Tafar Smith" }
]
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "sqlalchemy>=2.0",
  "asyncpg>=0.29",
  "alembic>=1.13",
  "redis>=5.0",
  "rq>=1.15",
  "httpx>=0.27",
  "beautifulsoup4>=4.12",
  "tenacity>=8.2",
  "structlog>=24.1",
  "python-dotenv>=1.0",
  "prometheus-client>=0.20",
  "google-auth>=2.30",
  "google-auth-oauthlib>=1.2",
  "google-api-python-client>=2.140"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "mypy>=1.8",
  "ruff>=0.4",
  "types-redis>=4.6"
]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
strict = true

